modify code as suggestions
