+++ paddle/pten/core/convert_utils.cc
@@ -12,31 +12,11 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License. */
 #include "paddle/pten/core/convert_utils.h"
-
+#include "paddle/pten/core/kernel_alias_name.h"
 // See Note [ Why still include the fluid headers? ]
 #include "paddle/fluid/platform/device/gpu/gpu_info.h"
 
 namespace pten {
-
-// the map between kernel name in fluid and pten
-const std::unordered_map<std::string, std::string> kernel_alias_name_map = {
-    {"reshape2", "reshape"},
-    {"fill_any_like", "full_like"},
-    {"fill_constant", "full"},
-    {"matmul_v2", "matmul"},
-    {"flatten_contiguous_range", "flatten"},
-    {"reduce_mean", "mean"},
-    {"reduce_sum", "sum"},
-    {"elementwise_add", "add"},
-    {"elementwise_sub", "subtract"},
-    {"elementwise_mul", "muliply"},
-    {"elementwise_div", "divide"},
-    // fluid kernel "mean/reshape/matmul/flatten/sum" should be deprecated
-    {"mean", "deprecated"},
-    {"reshape", "deprecated"},
-    {"matmul", "deprecated"},
-    {"flatten", "deprecated"},
-    {"sum", "deprecated"}};
 
 // TODO(chenweihang): Add other place trans cases later
 Backend TransToPtenBackend(const paddle::platform::Place& place) {
@@ -290,7 +270,7 @@ std::string DataType2String(DataType dtype) {
   }
 }
 
-std::string TransToPtenKernelName(const std::string& fluid_op_name) {
+const std::string& TransToPtenKernelName(const std::string& fluid_op_name) {
   if (kernel_alias_name_map.find(fluid_op_name) !=
       kernel_alias_name_map.end()) {
     return kernel_alias_name_map.at(fluid_op_name);
