+++ paddle/pten/core/convert_utils.h
@@ -32,7 +32,7 @@ namespace pten {
 using DataType = paddle::experimental::DataType;
 using DataLayout = paddle::experimental::DataLayout;
 
-std::string TransToPtenKernelName(const std::string& fluid_op_name);
+const std::string& TransToPtenKernelName(const std::string& fluid_op_name);
 
 Backend TransToPtenBackend(const paddle::platform::Place& place);
 DataType TransToPtenDataType(
