+++ paddle/pten/kernels/cuda/linalg.cu
@@ -71,7 +71,7 @@ PT_REGISTER_KERNEL("dot",
                    complex64,
                    complex128) {}
 
-PT_REGISTER_KERNEL("matmul",
+PT_REGISTER_KERNEL("matmul_v2",
                    CUDA,
                    ANY,
                    pten::Matmul,
